Extract sidebar components
