+++ app/javascript/home/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Layout from './components/Layout';
+import Home from './components/Home';
 
 const App = () => {
-  return <Layout />;
+  return <Home />;
 };
 
 export default App;
