+++ app/javascript/home/components/Home/index.tsx
@@ -1,0 +1,3 @@
+import { Home } from './Home';
+
+export default Home;
